Add files via upload
Procedure which adds two times together. The data type for integers is defined in the range 0..59. Hours, minutes, seconds are bundled as one record type time. 
A function that does the same as the above procedure.
The program passes the data and information flow analysis of Spark Ada and meaningful verification conditions are present.

This was developed by me as a solution to a High Integrity task from Swansea University.